save npy data as a default
